% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{add.cell.annotation}
\alias{add.cell.annotation}
\title{Add cell annotation columns to a cellinfo object}
\usage{
add.cell.annotation(so, cellinfo, vars, assay = NULL, overwrite = T)
}
\arguments{
\item{so}{A Seurat object.}

\item{cellinfo}{A cellinfo list object.}

\item{vars}{Character vector of metadata columns and/or gene names to add.}

\item{assay}{Assay to use for gene values. Defaults to \code{DefaultAssay(so)}.}

\item{overwrite}{Logical; if \code{FALSE}, only new annotation columns are added.}
}
\value{
Updated \code{cellinfo} list.
}
\description{
Imports metadata and/or gene expression values from a Seurat object into \code{cellinfo$cell.annotation} for the cells already represented in \code{cellinfo}.
}
\seealso{
\code{\link[=remove.cell.annotation]{remove.cell.annotation()}}, \code{\link[=refresh.cellinfo]{refresh.cellinfo()}}
}
