% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-subset.R
\name{removemarkers}
\alias{removemarkers}
\title{Remove markers from all marker groups}
\usage{
removemarkers(cellinfo, markers)
}
\arguments{
\item{cellinfo}{A cellinfo list object.}

\item{markers}{Character vector of genes to remove.}
}
\value{
Updated \code{cellinfo}.
}
\description{
Deletes specified genes from every vector in \code{cellinfo$markerlist} and refreshes derived fields.
}
